Dynamic-Fog-Nodes/fog-manager/requirements.txt
paho-mqtt==1.6.1
docker==6.1.3
requests==2.28.2
urllib3==1.26.15

